finished menus and Facades

+++ dev/DTOs/ItemDTO.java
@@ -2,16 +2,16 @@ package DTOs;
 
 public class ItemDTO {
     private String name;
-    private int weight;   // In whole numbers, in grams.
+    private double weight;   // In whole numbers, in grams.
     private Boolean condition;
 
     public ItemDTO() {}
-    public ItemDTO(String name, int weight, Boolean condition){
+    public ItemDTO(String name, double weight, Boolean condition){
         this.name = name;
         this.weight = weight;
         this.condition = condition;
     }
-    public ItemDTO(String name, int weight){
+    public ItemDTO(String name, double weight){
         this.name = name;
         this.weight = weight;
         this.condition = true;
@@ -19,8 +19,8 @@ public class ItemDTO {
 
     public String getName() {return name;}
     public void setName(String name) {this.name = name;}
-    public int getWeight() {return weight;}
-    public void setWeight(int weight) {this.weight = weight;}
+    public double getWeight() {return weight;}
+    public void setWeight(double weight) {this.weight = weight;}
     public Boolean getCondition() {return condition;}
     public void setCondition(Boolean condition) {this.condition = condition;}
